+++ app/src/main/java/com/x8bit/bitwarden/data/platform/datasource/network/di/PlatformNetworkModule.kt
@@ -7,7 +7,6 @@ import com.x8bit.bitwarden.data.platform.datasource.network.interceptor.BaseUrlI
 import com.x8bit.bitwarden.data.platform.datasource.network.interceptor.HeadersInterceptor
 import com.x8bit.bitwarden.data.platform.datasource.network.retrofit.Retrofits
 import com.x8bit.bitwarden.data.platform.datasource.network.retrofit.RetrofitsImpl
-import com.x8bit.bitwarden.data.platform.datasource.network.serializer.ZonedDateTimeSerializer
 import com.x8bit.bitwarden.data.platform.datasource.network.service.ConfigService
 import com.x8bit.bitwarden.data.platform.datasource.network.service.ConfigServiceImpl
 import com.x8bit.bitwarden.data.platform.datasource.network.service.EventService
@@ -23,8 +22,6 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import kotlinx.serialization.json.Json
-import kotlinx.serialization.modules.SerializersModule
-import kotlinx.serialization.modules.contextual
 import retrofit2.create
 import javax.inject.Singleton
 
@@ -103,23 +100,4 @@ object PlatformNetworkModule {
             sslManager = sslManager,
             json = json,
         )
-
-    @Provides
-    @Singleton
-    fun providesJson(): Json = Json {
-
-        // If there are keys returned by the server not modeled by a serializable class,
-        // ignore them.
-        // This makes additive server changes non-breaking.
-        ignoreUnknownKeys = true
-
-        // We allow for nullable values to have keys missing in the JSON response.
-        explicitNulls = false
-        serializersModule = SerializersModule {
-            contextual(ZonedDateTimeSerializer())
-        }
-
-        // Respect model default property values.
-        coerceInputValues = true
-    }
 }
