+++ app/src/test/java/com/x8bit/bitwarden/data/platform/datasource/disk/EventDiskSourceTest.kt
@@ -1,9 +1,9 @@
 package com.x8bit.bitwarden.data.platform.datasource.disk
 
+import com.bitwarden.core.di.CoreModule
 import com.x8bit.bitwarden.data.platform.base.FakeDispatcherManager
 import com.x8bit.bitwarden.data.platform.datasource.disk.dao.FakeOrganizationEventDao
 import com.x8bit.bitwarden.data.platform.datasource.disk.entity.OrganizationEventEntity
-import com.x8bit.bitwarden.data.platform.datasource.network.di.PlatformNetworkModule
 import com.x8bit.bitwarden.data.platform.datasource.network.model.OrganizationEventJson
 import com.x8bit.bitwarden.data.platform.manager.model.OrganizationEventType
 import kotlinx.coroutines.test.runTest
@@ -24,7 +24,7 @@ class EventDiskSourceTest {
 
     private val fakeOrganizationEventDao = FakeOrganizationEventDao()
     private val fakeDispatcherManager = FakeDispatcherManager()
-    private val json = PlatformNetworkModule.providesJson()
+    private val json = CoreModule.providesJson()
 
     private val eventDiskSource: EventDiskSource = EventDiskSourceImpl(
         organizationEventDao = fakeOrganizationEventDao,
