+++ app/src/test/java/com/x8bit/bitwarden/data/platform/datasource/disk/SettingsDiskSourceTest.kt
@@ -3,8 +3,8 @@ package com.x8bit.bitwarden.data.platform.datasource.disk
 import androidx.core.content.edit
 import app.cash.turbine.test
 import com.bitwarden.core.data.util.decodeFromStringOrNull
+import com.bitwarden.core.di.CoreModule
 import com.x8bit.bitwarden.data.platform.base.FakeSharedPreferences
-import com.x8bit.bitwarden.data.platform.datasource.network.di.PlatformNetworkModule
 import com.x8bit.bitwarden.data.platform.manager.model.AppResumeScreenData
 import com.x8bit.bitwarden.data.platform.repository.model.ClearClipboardFrequency
 import com.x8bit.bitwarden.data.platform.repository.model.UriMatchType
@@ -23,7 +23,7 @@ import java.time.Instant
 @Suppress("LargeClass")
 class SettingsDiskSourceTest {
     private val fakeSharedPreferences = FakeSharedPreferences()
-    private val json = PlatformNetworkModule.providesJson()
+    private val json = CoreModule.providesJson()
 
     private val settingsDiskSource = SettingsDiskSourceImpl(
         sharedPreferences = fakeSharedPreferences,
