+++ app/src/test/java/com/x8bit/bitwarden/data/platform/manager/PushManagerTest.kt
@@ -3,6 +3,7 @@ package com.x8bit.bitwarden.data.platform.manager
 import app.cash.turbine.test
 import com.bitwarden.core.data.util.asFailure
 import com.bitwarden.core.data.util.asSuccess
+import com.bitwarden.core.di.CoreModule
 import com.x8bit.bitwarden.data.auth.datasource.disk.AuthDiskSource
 import com.x8bit.bitwarden.data.auth.datasource.disk.model.AccountJson
 import com.x8bit.bitwarden.data.auth.datasource.disk.model.AccountTokensJson
@@ -12,7 +13,6 @@ import com.x8bit.bitwarden.data.platform.base.FakeDispatcherManager
 import com.x8bit.bitwarden.data.platform.base.FakeSharedPreferences
 import com.x8bit.bitwarden.data.platform.datasource.disk.PushDiskSource
 import com.x8bit.bitwarden.data.platform.datasource.disk.PushDiskSourceImpl
-import com.x8bit.bitwarden.data.platform.datasource.network.di.PlatformNetworkModule
 import com.x8bit.bitwarden.data.platform.datasource.network.model.PushTokenRequest
 import com.x8bit.bitwarden.data.platform.datasource.network.service.PushService
 import com.x8bit.bitwarden.data.platform.manager.dispatcher.DispatcherManager
@@ -63,7 +63,7 @@ class PushManagerTest {
             pushService = pushService,
             dispatcherManager = dispatcherManager,
             clock = clock,
-            json = PlatformNetworkModule.providesJson(),
+            json = CoreModule.providesJson(),
         )
     }
