+++ authenticator/src/test/java/com/bitwarden/authenticator/data/platform/datasource/disk/FeatureFlagDiskSourceTest.kt
@@ -4,8 +4,8 @@ import androidx.core.content.edit
 import app.cash.turbine.test
 import com.bitwarden.authenticator.data.platform.base.FakeSharedPreferences
 import com.bitwarden.authenticator.data.platform.datasource.disk.model.FeatureFlagsConfiguration
-import com.bitwarden.authenticator.data.platform.datasource.network.di.PlatformNetworkModule
 import com.bitwarden.authenticator.data.platform.manager.model.FlagKey
+import com.bitwarden.core.di.CoreModule
 import kotlinx.coroutines.test.runTest
 import kotlinx.serialization.json.JsonPrimitive
 import org.junit.jupiter.api.Assertions.assertEquals
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Test
 
 class FeatureFlagDiskSourceTest {
-    private val json = PlatformNetworkModule.providesJson()
+    private val json = CoreModule.providesJson()
 
     private val fakeSharedPreferences = FakeSharedPreferences()
 
@@ -58,7 +58,7 @@ class FeatureFlagDiskSourceTest {
 private const val FEATURE_FLAGS_CONFIGURATION_JSON = """
 {
     "featureFlags" : {
-        "bitwarden-authentication-enabled" : true    
+        "bitwarden-authentication-enabled" : true
     }
 }
