+++ core/src/main/kotlin/com/bitwarden/core/di/CoreModule.kt
@@ -1,0 +1,38 @@
+package com.bitwarden.core.di
+
+import com.bitwarden.core.data.serializer.ZonedDateTimeSerializer
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.modules.SerializersModule
+import kotlinx.serialization.modules.contextual
+import javax.inject.Singleton
+
+/**
+ * Provides dependencies from the core module.
+ */
+@Module
+@InstallIn(SingletonComponent::class)
+object CoreModule {
+
+    @Provides
+    @Singleton
+    fun providesJson(): Json = Json {
+
+        // If there are keys returned by the server not modeled by a serializable class,
+        // ignore them.
+        // This makes additive server changes non-breaking.
+        ignoreUnknownKeys = true
+
+        // We allow for nullable values to have keys missing in the JSON response.
+        explicitNulls = false
+        serializersModule = SerializersModule {
+            contextual(ZonedDateTimeSerializer())
+        }
+
+        // Respect model default property values.
+        coerceInputValues = true
+    }
+}
